can't get it to work
I have tried for hours to get this to work. It looks like I am doing something similar to the favorites, but I just don't understand it. I really tried. I could essentially just copy what is on the favorites and get it to work but at this point I really am just hating swift and xcode because it doesn't make any sense